requirements.txt
phi-data
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
google-generativeai
